Dialogs Improvement

+++ composeApp/src/commonMain/kotlin/sharedData/UserDeviceData.kt
@@ -9,13 +9,13 @@ expect fun getDeviceMake(): String      //device manufacturer
 expect fun getDeviceId(): String        //temporary random app ID
 
 @Composable
-expect fun actualWidthFactor(): Float        //real display width in Dp
+expect fun actualWidthFactor(): Int        //real display width in Dp
 
 @Composable
 expect fun getScreenWidth(): Int
 
 @Composable
-expect fun actualHeightFactor(): Float       //real display height in Dp
+expect fun actualHeightFactor(): Int       //real display height in Dp
 
 @Composable
 expect fun getScreenHeight(): Int
